2021.09.09

+++ controller/tweet.js
@@ -1,42 +1,52 @@
 import * as tweetRepository from '../data/tweet.js';
 
-export async function getTweets(request, response) {
-    const username = request.query.username;
-    const data = await (username 
-        ? tweetRepository.getAllByUserName(username)
-        : tweetRepository.getAll()); // 여기는 이거 자체가 promise니까,,
-    response.status(200).json(data);
+export async function getTweets(req, res) {
+  const username = req.query.username;
+  const data = await (username
+    ? tweetRepository.getAllByUsername(username)
+    : tweetRepository.getAll());
+  res.status(200).json(data);
 }
 
-export async function getTweet(request, response) {
-    const id = request.params.id;
-    const tweet = await tweetRepository.getById(id);
-    if(tweet) {
-        response.sendStatus(200).json(tweet);
-    } else {
-        response.status(404).json({message: `Tweet id(${id}) not found`});
-    }
+export async function getTweet(req, res, next) {
+  const id = req.params.id;
+  const tweet = await tweetRepository.getById(id);
+  if (tweet) {
+    res.status(200).json(tweet);
+  } else {
+    res.status(404).json({ message: `Tweet id(${id}) not found` });
+  }
 }
 
-export async function createTweet(request, response) {
-    const {text, name, username} = request.body;
-    const tweet = await tweetRepository.create(text, name, username);
-    response.status(201).json(tweet);
+export async function createTweet(req, res, next) {
+  const { text } = req.body;
+  const tweet = await tweetRepository.create(text, req.userId);
+  res.status(201).json(tweet);
 }
 
-export async function updateTweet(request, response) {
-    const id = request.params.id;
-    const text = request.body.text;
-    const tweet = await tweetRepository.update(id, text);
-    if(tweet) {
-        response.status(200).json(tweet);
-    } else {
-        response.status(404).json({message: `Tweet id(${id}) not found`});
-    }
+export async function updateTweet(req, res, next) {
+  const id = req.params.id;
+  const text = req.body.text;
+  const tweet = await tweetRepository.getById(id);
+  if (!tweet) {
+    return res.status(404).json({ message: `Tweet not found: ${id}` });
+  }
+  if (tweet.userId !== req.userId) {
+    return res.sendStatus(403);
+  }
+  const updated = await tweetRepository.update(id, text);
+  res.status(200).json(updated);
 }
 
-export async function deleteTweet(request, response) {
-    const id = request.params.id;
-    await tweetRepository.remove(id);
-    response.send(204);
+export async function deleteTweet(req, res, next) {
+  const id = req.params.id;
+  const tweet = await tweetRepository.getById(id);
+  if (!tweet) {
+    return res.status(404).json({ message: `Tweet not found: ${id}` });
+  }
+  if (tweet.userId !== req.userId) {
+    return res.sendStatus(403);
+  }
+  await tweetRepository.remove(id);
+  res.sendStatus(204);
 }
